fixed z index and started data process
